src/requirements.txt
numpy==1.11.0
h5py==2.7.0
tensorflow==1.0.1
Pillow==5.0.0
